Fix float16 action-mask logit-masking NaN (mixed precision)
When PR makes action masks float16, the masking term `-1e8 * (1 - mask)` is evaluated
in float16: `-1e8` overflows to `-inf` for masked actions, and `inf * 0` then turns the
*valid*-action logits into `NaN`. That poisons the entropy/gradient and collapses training
(observed on a transformer USA100 run: blocking 0.045 -> 0.858 vs the float32 baseline).

Cast the mask up to float32 inside the masking expression at all 10 sites in ppo.py and
train_utils.py, so the `-1e8` constant stays finite. This is a no-op in default (float32-mask)
mode -- bit-identical to before -- and makes the float16-mask path match it.

Verified: forcing the float tiers back to float32 (`--small_float_dtype=float32
--time_dtype=float32`) already trains bit-identically to the baseline, and int16/int8 are
bit-identical too; the divergence was solely this float16-mask overflow. Adds a regression
test (mixed_precision_test.py::MaskLogitMaskingTest) pinning the hazard and the fix.

The multidevice variants already use the safe `jnp.where(mask, logits, -1e8)` form (no
arithmetic overflow), so they need no change.

+++ xlron/train/ppo.py
@@ -455,9 +455,9 @@ def _loss_fn(
     if config.env_type.lower() == "vone":
         # VONE: source, path, destination actions
         vone_batch = cast(VONETransition, traj_batch)
-        pi_source = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_s)))
-        pi_path = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_p)))
-        pi_dest = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_d)))
+        pi_source = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_s.astype(jnp.float32))))
+        pi_path = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_p.astype(jnp.float32))))
+        pi_dest = distrax.Categorical(logits=pi._logits + (-1e8 * (1 - vone_batch.action_mask_d.astype(jnp.float32))))
         action_s = traj_batch.action[:, 0]
         action_p = traj_batch.action[:, 1]
         action_d = traj_batch.action[:, 2]
@@ -476,7 +476,7 @@ def _loss_fn(
 
         if config.GNN_OUTPUT_RSA:
             pi_masked = distrax.Categorical(
-                logits=path_dist._logits + (-1e8 * (1 - traj_batch.action_mask))
+                logits=path_dist._logits + (-1e8 * (1 - traj_batch.action_mask.astype(jnp.float32)))
             )
             path_log_prob = pi_masked.log_prob(path_actions)
             path_entropy = pi_masked.entropy()
@@ -514,7 +514,7 @@ def _loss_fn(
     else:
         # Standard action masking
         pi_masked = distrax.Categorical(
-            logits=pi[0]._logits + (-1e8 * (1 - traj_batch.action_mask))
+            logits=pi[0]._logits + (-1e8 * (1 - traj_batch.action_mask.astype(jnp.float32)))
         )
         # Ratio will be policy/masked_policy - also known as off-policy invalid action masking
         log_prob = (
